Create SECURITY.md
Report suspicious transactions
Core team network
After greeting
As pioneers in the Pi Network project and we believe in prioritizing the 
public interest over private interest,
After checking and verifying, we members of one of the largest Arab 
groups on one of the social networking applications 
Telegram(https://t.me/pinetwork111122*, found suspicious operations under 
the guise of the name Bay Network, carried out by people managing this 
group, which includes more than 9,000 members . , who are being exploited 
to obtain coverage. These operations have been verified and are as 
follows:
Money laundering through stolen intercontinental currencies that are 
stolen from pioneers all over the world by criminal groups in Asian 
countries in cooperation with parties in this group and the impact of 
this on pioneers in the world. Then these balances are transferred to the 
Arab Republic of Egypt, specifically to the wallets of the officials of 
this group mentioned above, and are sold through these names. The matter 
does not stop there, but rather this process is carried out to obtain 
cash liquidity from these currencies after selling the currencies via the 
application (CoinSkro) (@Hanox80) You can track the operations that take 
place on this application to investigate and follow up on the currencies 
stolen on it and their size and convert them into dollars and then begin 
the process of supporting entities and individuals in other countries by 
transferring this money to them, which of course are areas witnessing 
armed conflicts such as Yemen and Palestine, and dangerous groups that 
have been criminalized internationally
Below are attached to you the names of those involved in more than one 
country,Egypt, Palestine, and other
countries
Username in Pi Network
@wistam 
( Palestine)
—-
@naderlopopolo ( Egypt)
——
@metto1981 ( Egypt)
——-
@Thescorpion2020 ( Egypt)
——
madeeh muhsin (Egypt)
telephone number 
+201069514141